% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-glmmsel.R
\name{coef.cv.glmmsel}
\alias{coef.cv.glmmsel}
\title{Coefficient function for \code{cv.glmmsel} object}
\usage{
\method{coef}{cv.glmmsel}(object, lambda = "lambda.min", ...)
}
\arguments{
\item{object}{an object of class \code{cv.glmmsel}}

\item{lambda}{a value of the regularisation parameter}

\item{...}{any other arguments}
}
\value{
An array of coefficients.
}
\description{
Extract cluster coefficients for a cross-validated value of the regularisation
parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
