% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmsel.R
\name{coef.glmmsel}
\alias{coef.glmmsel}
\title{Coefficient function for \code{glmmsel} object}
\usage{
\method{coef}{glmmsel}(object, lambda = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{glmmsel}}

\item{lambda}{a value of the regularisation parameter}

\item{...}{any other arguments}
}
\value{
An array of coefficients.
}
\description{
Extracts coefficients for a specified value of the regularisation parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
