% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmsel.R
\name{fixef.glmmsel}
\alias{fixef.glmmsel}
\title{Fixed effects function for \code{glmmsel} object}
\usage{
\method{fixef}{glmmsel}(object, lambda = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{glmmsel}}

\item{lambda}{a value of the regularisation parameter}

\item{...}{any other arguments}
}
\value{
A matrix of fixed effects.
}
\description{
Extracts fixed effects for a specified value of the regularisation parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
