% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmsel-package.R
\name{fixef}
\alias{fixef}
\title{Fixed effects function}
\usage{
fixef(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{any other arguments}
}
\value{
Depends on the specific method implementation.
}
\description{
Generic function for extracting fixed effects from model objects.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
