% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-glmmsel.R
\name{plot.cv.glmmsel}
\alias{plot.cv.glmmsel}
\title{Plot function for \code{cv.glmmsel} object}
\usage{
\method{plot}{cv.glmmsel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{cv.glmmsel}}

\item{...}{any other arguments}
}
\value{
A plot of the cross-validation results.
}
\description{
Plot the cross-validation loss as a function of the regularisation parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
