% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmsel.R
\name{plot.glmmsel}
\alias{plot.glmmsel}
\title{Plot function for \code{glmmsel} object}
\usage{
\method{plot}{glmmsel}(x, cluster = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{glmmsel}}

\item{cluster}{the cluster whose coefficients to plot}

\item{...}{any other arguments}
}
\value{
A plot of the coefficient profiles.
}
\description{
Plots the coefficients as a function of the regularisation parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
