% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-glmmsel.R
\name{predict.cv.glmmsel}
\alias{predict.cv.glmmsel}
\title{Predict function for \code{cv.glmmsel} object}
\usage{
\method{predict}{cv.glmmsel}(object, x.new, cluster.new, lambda = "lambda.min", ...)
}
\arguments{
\item{object}{an object of class \code{cv.glmmsel}}

\item{x.new}{a matrix of new values for the predictors}

\item{cluster.new}{a vector identifying the clusters that the rows of \code{x.new} belong to}

\item{lambda}{a value of the regularisation parameter}

\item{...}{any other arguments}
}
\value{
A matrix of predictions.
}
\description{
Generates predictions for new data using a cross-validated value of the
regularisation parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
