% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-glmmsel.R
\name{ranef.cv.glmmsel}
\alias{ranef.cv.glmmsel}
\title{Random effects function for \code{cv.glmmsel} object}
\usage{
\method{ranef}{cv.glmmsel}(object, lambda = "lambda.min", ...)
}
\arguments{
\item{object}{an object of class \code{cv.glmmsel}}

\item{lambda}{a value of the regularisation parameter}

\item{...}{any other arguments}
}
\value{
A matrix of random effects.
}
\description{
Extract random effects for a cross-validated value of the regularisation parameter.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
